requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
sqlalchemy>=1.4.0
mysql-connector-python>=8.0.0
watchdog>=2.1.0
pyinstaller>=5.0.0
